[web] Rename file `page` -> `BlogPage`
