+++ apps/web/pages/w/[[...slug]].tsx
@@ -5,7 +5,7 @@ import {
   buildGetStaticPaths,
   buildGetStaticProps,
   type BlogPageProps,
-} from '../../posts/components/page';
+} from '../../posts/components/BlogPage';
 import { type PostCategoryType } from '../../posts/lib/get-posts';
 
 const POST_CATEGORY_TYPE: PostCategoryType = 'blog';
